Suppressed a warning

+++ PulsingHalo/MultiplePulsingHaloLayer.m
@@ -75,7 +75,7 @@
             [self addSublayer:layer];
         });
     }
-    NSLog(@"sub count:%i", self.sublayers.count);
+    NSLog(@"sub count:%lu", (unsigned long)self.sublayers.count);
 }
 
 - (void)dealloc {
